+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYRFKV0PXC13BPRZ1C1PSVX7"
+        "01KYRRWH0TB1JG4C68JSBEF8B5"
     )
     r"""The unique identifier of the data source"""