+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -507,9 +507,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-30T03:03:04.599Z"
+    created: Optional[str] = "2026-07-30T05:45:05.932Z"
 
-    updated: Optional[str] = "2026-07-30T03:03:04.599Z"
+    updated: Optional[str] = "2026-07-30T05:45:05.932Z"
 
     updated_by_id: OptionalNullable[str] = UNSET