+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYRFKV0NDFK6WA55WAKM6AD1"
+        "01KYRRWH0SQPWR2WMXSC1386VN"
     )
     r"""The unique identifier of the data source"""