+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYRFKV0REV15ANM5M40FPTDC"
+        "01KYRRWH0WC5J58R8ZHEKRRFAQ"
     )
     r"""The unique identifier of the data source"""