+++ packages/orq-rc/RELEASES.md
@@ -9839,3 +9839,13 @@ Based on:
 - [python v4.13.0-rc.17] packages/orq-rc
 ### Releases
 - [PyPI v4.13.0-rc.17] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.17 - packages/orq-rc
+
+## 2026-07-30 05:48:16
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.18] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.18] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.18 - packages/orq-rc